+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/cti/catalog/client/SnapshotClient.java
@@ -20,6 +20,9 @@ import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpResponse;
 import org.apache.hc.client5.http.impl.classic.HttpClients;
+import org.apache.hc.core5.http.Header;
+import org.apache.hc.core5.http.HttpHeaders;
+import org.apache.hc.core5.http.message.BasicHeader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.env.Environment;
@@ -33,6 +36,9 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
+import java.util.List;
+
+import com.wazuh.contentmanager.settings.PluginSettings;
 
 /** Client responsible for downloading CTI snapshots from a remote source. */
 public class SnapshotClient {
@@ -58,7 +64,11 @@ public class SnapshotClient {
      * @throws URISyntaxException If the provided URI is invalid.
      */
     public Path downloadFile(String snapshotURI) throws IOException, URISyntaxException {
-        try (CloseableHttpClient client = HttpClients.createDefault()) {
+        List<Header> defaultHeaders =
+                List.of(
+                        new BasicHeader(HttpHeaders.USER_AGENT, PluginSettings.getInstance().getUserAgent()));
+        try (CloseableHttpClient client =
+                HttpClients.custom().setDefaultHeaders(defaultHeaders).build()) {
             // Setup
             final URI uri = new URI(snapshotURI);
             final HttpGet request = new HttpGet(uri);